refactor : 일반 모달에 연결 되어 있던 신고하기 모달 훅 따로 빼기
